Crate runtime_eden

Modules

api

constants

Structs

CheckInherents

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaId

Unique identifier of a parachain.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SessionKeys

Enums

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

VERSION

This runtime version. This should not be thought of as classic Semver (major/minor/tiny). This triplet have different semantics and mis-interpretation could cause problems. In particular: bug fixes should result in an increment of spec_version and possibly authoring_version, absolutely not impl_version since they change the semantics of the runtime.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Allocations

AllocationsOracles

AllocationsOraclesConfig

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

CollatorSelection

CollatorSelectionConfig

CompanyReserve

CompanyReserveConfig

Contracts

CumulusXcm

DaoReserve

DaoReserveConfig

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

Header

Block header type as expected by this runtime.

Identity

InternationalReserve

InternationalReserveConfig

Mandate

MessageQueue

Migrations

Multisig

NodleUniques

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Preimage

Proxy

RandomnessCollectiveFlip

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sponsorship

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Uniques

UsaReserve

UsaReserveConfig

Utility

Vesting

VestingConfig

XTokens

XcmpQueue