runtime_eden
Module constants

Structs

Perbill

A fixed point representation of a number in the range [0, 1].

Perquintill

A fixed point representation of a number in the range [0, 1].

RuntimeBlockWeights

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~5% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

BLOCK_PROCESSING_VELOCITY

How many parachain blocks are processed by the relay chain per parent. Limits the number of blocks authored per slot.

CONTRACTS_DEBUG_OUTPUT

DAYS

DAYS_RELAY_CHAIN

EPOCH_DURATION_IN_BLOCKS

Time and blocks.

EPOCH_DURATION_IN_SLOTS

EXISTENTIAL_DEPOSIT

HOURS

HOURS_RELAY_CHAIN

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 6 second average block time.

MICRO_NODL

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLI_NODL

MINUTES

MINUTES_RELAY_CHAIN

NANO_NODL

NODL

Money matters.

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

POLKADOT_CENT

POLKADOT_EXISTENTIAL_DEPOSIT

PRIMARY_PROBABILITY

RELAY_CHAIN_SLOT_DURATION_MILLIS

Relay chain slot duration, in milliseconds.

SLOT_DURATION

TARGET_BLOCK_FULLNESS

Fee-related. The block saturation level. Fees will be updates based on this value.

UNINCLUDED_SEGMENT_CAPACITY

Maximum number of blocks simultaneously accepted by the Runtime, not yet included into the relay chain.

Functions

deposit