Type Alias runtime_eden::CheckedExtrinsic

pub type CheckedExtrinsic = CheckedExtrinsic<AccountId, RuntimeCall, SignedExtra>;

Extrinsic type that has already been checked.

Aliased Type

struct CheckedExtrinsic {
    pub signed: Option<(AccountId32, (CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>, ChargeSponsor<Runtime>))>,
    pub function: RuntimeCall,
}

Fields
signed: Option<(AccountId32, (CheckSpecVersion<Runtime>, CheckTxVersion<Runtime>, CheckGenesis<Runtime>, CheckMortality<Runtime>, CheckNonce<Runtime>, CheckWeight<Runtime>, ChargeTransactionPayment<Runtime>, ChargeSponsor<Runtime>))>

Who this purports to be from and the number of extrinsics have come before from the same signer, if anyone (note this is not a signature).

function: RuntimeCall

The function that should be called.