Crate runtime_eden

Modules

api

constants

Structs

GenesisConfig

Origin

The runtime origin type represanting the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaId

Unique identifier of a parachain.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

SessionKeys

Enums

Call

Event

OriginCaller

Constants

VERSION

This runtime version. This should not be thought of as classic Semver (major/minor/tiny). This triplet have different semantics and mis-interpretation could cause problems. In particular: bug fixes should result in an increment of spec_version and possibly authoring_version, absolutely not impl_version since they change the semantics of the runtime.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with SKIP_WASM_BUILD, the function panics.

Type Definitions

Address

The address format for describing accounts.

AllModulesDeprecated

All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystemDeprecated

All modules included in the runtime as a nested tuple of types.

AllPallets

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

Allocations

AllocationsOracles

AllocationsOraclesConfig

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

CompanyReserve

CompanyReserveConfig

CumulusXcm

EmergencyShutdown

Executive

Header

Block header type as expected by this runtime.

InternationalReserve

InternationalReserveConfig

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

Poa

RandomnessCollectiveFlip

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UsaReserve

UsaReserveConfig

Utility

ValidatorsSet

ValidatorsSetConfig

Vesting

VestingConfig