Module runtime_eden::constants

Structs

Perbill

A fixed point representation of a number in the range [0, 1].

Perquintill

A fixed point representation of a number in the range [0, 1].

RuntimeBlockWeights

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~10% of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

CONTRACTS_DEBUG_OUTPUT

DAYS

DAYS_RELAY_CHAIN

EPOCH_DURATION_IN_BLOCKS

EPOCH_DURATION_IN_SLOTS

EXISTENTIAL_DEPOSIT

HOURS

HOURS_RELAY_CHAIN

MAXIMUM_BLOCK_WEIGHT

We allow for .5 seconds of compute with a 16 second average block time.

MICRO_NODL

MILLISECS_PER_BLOCK

Time and blocks.

MILLISECS_PER_RELAY_CHAIN_BLOCK

MILLI_NODL

MINUTES

MINUTES_RELAY_CHAIN

NANO_NODL

NODL

Money matters.

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

PRIMARY_PROBABILITY

SLOT_DURATION

TARGET_BLOCK_FULLNESS

Fee-related. The block saturation level. Fees will be updates based on this value.

Functions

deposit